+++ examples/fitness/lib/main.dart
@@ -5,6 +5,7 @@
 library fitness;
 
 import 'package:playfair/playfair.dart' as playfair;
+import 'package:sky/animation.dart';
 import 'package:sky/painting/text_style.dart';
 import 'package:sky/theme/colors.dart' as colors;
 import 'package:sky/widgets.dart';
