+++ sky/packages/sky/lib/animation.dart
@@ -1,0 +1,19 @@
+// Copyright 2015 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+/// The Sky animation engine.
+///
+/// This library includes and re-exports all Sky animation classes.
+library animation;
+
+export 'package:sky/src/rendering/auto_layout.dart';
+
+export 'src/animation/animated_simulation.dart';
+export 'src/animation/animated_value.dart';
+export 'src/animation/animation_performance.dart';
+export 'src/animation/curves.dart';
+export 'src/animation/forces.dart';
+export 'src/animation/scheduler.dart';
+export 'src/animation/scroll_behavior.dart';
+export 'src/animation/timeline.dart';
