--- sky/packages/sky/lib/animation/direction.dart
@@ -1,8 +1,0 @@
-// Copyright 2015 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-enum Direction {
-  forward,
-  reverse
-}
