+++ sky/packages/sky/lib/painting/README.md
@@ -1,3 +1,3 @@
 This directory contains painting-related libraries that only depend on
-core Dart libraries, ../base/*, and ../animation/*. Note that
-../animation/* depends on the Newton Dart library also.
+core Dart libraries, `../base/*`, and [animation.dart](../animation.dart). Note
+that [animation.dart](../animation.dart) depends on the Newton Dart library.
