+++ sky/packages/sky/lib/src/animation/README.md
@@ -1,2 +1,2 @@
 This directory contains animation-related libraries that only depend
-on core Dart libraries, the Newton Dart library, and ../base/*.
+on core Dart libraries and the Newton Dart library.
