+++ sky/packages/sky/lib/src/animation/forces.dart
@@ -6,22 +6,25 @@ import 'package:newton/newton.dart';
 
 // Base class for creating Simulations for the animation Timeline.
 abstract class Force {
+  const Force();
+
   Simulation release(double position, double velocity);
 }
 
 class SpringForce extends Force {
-  SpringForce(this.spring, { this.left: 0.0, this.right: 1.0 });
+  const SpringForce(this.spring, { this.left: 0.0, this.right: 1.0 });
 
   final SpringDescription spring;
 
   // Where to put the spring's resting point when releasing left or right,
   // respectively.
-  final double left, right;
+  final double left;
+  final double right;
 
   // We overshoot the target by this distance, but stop the simulation when
   // the spring gets within this distance (regardless of how fast it's moving).
   // This causes the spring to settle a bit faster than it otherwise would.
-  static final Tolerance tolerance = new Tolerance(
+  static const Tolerance tolerance = const Tolerance(
     velocity: double.INFINITY,
     distance: 0.01
   );
