+++ sky/packages/sky/lib/src/widgets/transitions.dart
@@ -2,14 +2,13 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:sky/animation/animated_value.dart';
-import 'package:sky/animation/animation_performance.dart';
+import 'package:sky/animation.dart';
 import 'package:sky/src/widgets/animated_component.dart';
 import 'package:sky/src/widgets/basic.dart';
 import 'package:sky/src/widgets/framework.dart';
 import 'package:vector_math/vector_math.dart';
 
-export 'package:sky/animation/direction.dart' show Direction;
+export 'package:sky/animation.dart' show Direction;
 
 // A helper class to anchor widgets to one another. Pass an instance of this to
 // a Transition, then use the build() method to create a child with the same
