+++ sky/unit/test/widget/widget_tester.dart
@@ -1,8 +1,8 @@
 import 'dart:sky' as sky;
 
+import 'package:sky/animation.dart';
 import 'package:sky/rendering.dart';
 import 'package:sky/widgets.dart';
-import 'package:sky/base/scheduler.dart' as scheduler;
 
 import '../engine/mock_events.dart';
